Updating the rendering system. Starting on looking at "config" files maybe for tiles, items, and settings. Building up some sort of idea to move on with this framework to.

+++ src/main/java/inferno/saigo/client/rendering/ObjectRenderingCoord.java
@@ -1,18 +1,18 @@
 package inferno.saigo.client.rendering;
 
 public abstract class ObjectRenderingCoord extends ObjectRendering {
-    private final int x, y;
+    private final float x, y;
 
-    public ObjectRenderingCoord(int x, int y) {
+    public ObjectRenderingCoord(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 }
